+++ apps/ledger-live-desktop/src/renderer/analytics/segment.ts
@@ -70,7 +70,7 @@ const getMarketWidgetAnalytics = (state: State) => {
 
   return {
     hasMarketWidget: !marketWidget?.enabled ? "Null" : hasMarketWidgetActivated ? "Yes" : "No",
-    hasSeenWidgetImprovements: marketWidget?.params?.enableNewFeature ? "Yes" : "No",
+    hasMarketWidgetV2: marketWidget?.params?.enableNewFeature ? "Yes" : "No",
   };
 };