+++ apps/ledger-live-mobile/src/newArch/features/Accounts/Navigator.tsx
@@ -2,7 +2,7 @@ import React, { useCallback, useMemo } from "react";
 import { Platform } from "react-native";
 import { createStackNavigator } from "@react-navigation/stack";
 import { useTheme } from "styled-components/native";
-import { useRoute } from "@react-navigation/native";
+import { useNavigation, useRoute } from "@react-navigation/native";
 import { ScreenName } from "~/const";
 import { getStackNavigatorConfig } from "~/navigation/navigatorConfig";
 import { track } from "~/analytics";
@@ -13,11 +13,13 @@ import ScanDeviceAccounts from "LLM/features/Accounts/screens/ScanDeviceAccounts
 import { AccountsListNavigator } from "./screens/AccountsList/types";
 import { useFeature } from "@ledgerhq/live-common/featureFlags/index";
 import AccountsList from "LLM/features/Accounts/screens/AccountsList";
+import { NavigationHeaderBackButton } from "~/components/NavigationHeaderBackButton";
 
 export default function Navigator() {
   const { colors } = useTheme();
   const route = useRoute();
   const accountListUIFF = useFeature("llmAccountListUI");
+  const navigation = useNavigation();
 
   const onClose = useCallback(() => {
     track("button_clicked", {
@@ -34,6 +36,14 @@ export default function Navigator() {
     [colors, onClose],
   );
 
+  const onPressBack = useCallback(() => {
+    track("button_clicked", {
+      button: "Back",
+      page: route.name,
+    });
+    navigation.goBack();
+  }, [route, navigation]);
+
   return (
     <Stack.Navigator
       screenOptions={{
@@ -48,6 +58,7 @@ export default function Navigator() {
         options={{
           headerTitle: "",
         }}
+        initialParams={route.params}
       />
 
       {/* Scan accounts from device */}
@@ -64,6 +75,7 @@ export default function Navigator() {
           component={AccountsList}
           options={{
             headerTitle: "",
+            headerLeft: () => <NavigationHeaderBackButton onPress={onPressBack} />,
           }}
         />
       )}