+++ apps/ledger-live-mobile/src/newArch/features/Accounts/screens/AccountsList/types.ts
@@ -1,10 +1,12 @@
+import { Account, TokenAccount } from "@ledgerhq/types-live";
 import { ScreenName } from "~/const";
 
 export type AccountsListNavigator = {
   [ScreenName.AccountsList]: {
     sourceScreenName: ScreenName;
-    showHeader?: boolean;
-    canAddAccount?: boolean;
-    isSyncEnabled?: boolean;
+    showHeader?: boolean | string;
+    canAddAccount?: boolean | string;
+    isSyncEnabled?: boolean | string;
+    specificAccounts?: Account[] | TokenAccount[];
   };
 };